Move serialize and deserialize into storage submodule
