+++ tct/src/internal/frontier/item/builder.rs
@@ -1,6 +1,9 @@
 use super::*;
 
-use deserialize::{IResult, Instruction, Unexpected};
+use crate::storage::{
+    deserialize::{IResult, Unexpected},
+    Instruction,
+};
 
 /// A builder for the tip of the frontier.
 pub struct Builder {
