+++ tct/src/internal/frontier/tier/builder.rs
@@ -1,6 +1,9 @@
 use super::*;
 
-use deserialize::{IResult, Instruction, Unexpected};
+use crate::storage::{
+    deserialize::{IResult, Unexpected},
+    Instruction,
+};
 
 /// A builder for a frontier tier.
 pub struct Builder<Item: Built + Focus>
